+++ resources/js/components/TipTapSource.vue
@@ -5,6 +5,7 @@
     text
     @click="openModal"
     icon="fa-solid fa-code"
+    data-test="tip-tap-source-button"
   />
   <Dialog
     v-model:visible="modalVisible"
@@ -13,6 +14,7 @@
     :style="{ width: '500px' }"
     :breakpoints="{ '575px': '90vw' }"
     :draggable="false"
+    data-test="tip-tap-source-dialog"
   >
 
     <Textarea
@@ -20,6 +22,7 @@
       v-model="source"
       class="w-full mt-2"
       rows="5"
+      data-test="source-textarea"
     />
 
     <template #footer>
@@ -28,10 +31,12 @@
             severity="secondary"
             @click="modalVisible = false"
             :label="$t('app.cancel')"
+            data-test="dialog-cancel-button"
           />
           <Button
             @click="save"
             :label="$t('app.save')"
+            data-test="dialog-save-button"
           />
         </div>
     </template>